Exit worker threads immediately when their input channels disconnect
This stops the worker threads from sitting in a tight busy loop
consuming 100% CPU if the channels they listen for input on are ever
dropped.

This can for example happen if the thread owning the sender panics. Even
if the panic does result in the process eventually aborting, this stops
the worker threads to go crazy while the shut down happens.